修正距离运算可能溢出的问题，修正盟友判断中this可能为空的问题，删除SelectWeapon中的无用代码

+++ YRpp/YRMathVector.h
@@ -81,7 +81,7 @@ public:
 	//magnitude squared
 	double MagnitudeSquared() const
 	{
-		return *this * *this;
+		return static_cast<double>(*this) * static_cast<double>(*this);
 	}
 	//distance from another vector
 	double DistanceFrom(const Vector2D& a) const
@@ -202,7 +202,7 @@ public:
 	//magnitude squared
 	double MagnitudeSquared() const
 	{
-		return *this * *this;
+		return static_cast<double>(*this) * static_cast<double>(*this);
 	}
 	//distance from another vector
 	double DistanceFrom(const Vector3D& a) const
